requirements.txt
pandas
flask_cors==5.0.0